# docs/swagger.yaml
basePath: /
definitions:
  models.Subscription:
    properties:
      endDate:
        type: string
      id:
        format: int64
        type: integer
      price:
        type: integer
      serviceName:
        type: string
      startDate:
        type: string
      userID:
        type: string
    type: object
host: localhost:8090
info:
  contact:
    name: Алексей
    url: https://github.com/EvilCommunist
  description: REST сервис для управления подписками пользователей.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Возвращает список подписок пользователя. Фильтрация по названию
        сервиса – опциональна.
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "400":
          description: Не указан user_id или неверный формат
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties: true
            type: object
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт запись о подписке. Даты передаются в формате MM-YYYY.
      parameters:
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties: true
            type: object
      summary: Создать новую подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку по ID.
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties: true
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Возвращает полную информацию о подписке.
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Обновляет одно или несколько полей (price, start_date, end_date).
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties: true
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/totalcost:
    get:
      consumes:
      - application/json
      description: Возвращает стоимость подписок пользователя за указанный период.
        Фильтрация по названию сервиса – опциональна.
      parameters:
      - description: Начало просматриваемого периода (MM-YYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: Конец просматриваемого периода (MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      - description: UUID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: total_amount
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "400":
          description: Не указан user_id или неверный формат
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties: true
            type: object
      summary: Получить стоимость всех подписок за этот период
      tags:
      - subscriptions totalcost
schemes:
- http
swagger: "2.0"